+++ libra_toolbox/neutron_detection/activation_foils/calibration.py
@@ -1,6 +1,35 @@
 import datetime
 
-from libra_toolbox.neutron_detection.activation_foils import CheckSource
+
+class CheckSource:
+    nuclide: str
+    energy: list
+    intensity: list
+    activity_date: datetime.date
+    activity: float
+    half_life: float
+
+    def __init__(
+        self,
+        nuclide: str,
+        energy: list,
+        intensity: list,
+        activity_date: datetime.date,
+        activity: float,
+        half_life: float,
+    ):
+        self.nuclide = nuclide
+        self.energy = energy
+        self.intensity = intensity
+        self.activity_date = activity_date
+        self.activity = activity
+        self.half_life = half_life
+
+
+class ActivationFoil:
+    nuclide: str
+    mass: float
+
 
 uCi_to_Bq = 3.7e4
